// src/api/generated/model/signUpBody.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Bistro App Backend API
 * OpenAPI contract for generated frontend API clients.
 * OpenAPI spec version: 0.1.0
 */

export interface SignUpBody {
  email: string;
  /** @minLength 6 */
  password: string;
  /** @minLength 1 */
  name: string;
}